Use indexOf instead of raw loop
There are a couple of places where removeFromArray was implemented using a raw loop instead of the existing indexOf method that javascript has.  One of these areas is in demo code that others will copy and paste.  We should try to have our code reuse algorithms that already exist because it is not only simpler code but also sets a good example for others who work with skyux2.