Update file attachment demo acceptedTypes type (#593)
* Update file attachment demo acceptedTypes type

The acceptedTypes paramter on the file attachments component is described as a comma delimited string of mime file types, not a string array. Updating the sample to reflect this.

* Updating template to use acceptedTypes variable

Demo file attachments component now makes use of the acceptedTypes parameter to demonstrate only allowing jpeg and png type files.

* Fixed mistake where the wrong attribute was used

Accidentally left in the wrong attribute - 'allowedTypes' should be 'acceptedTypes'. Sorry!